Add IMU pose offset relative to base
Added a simple pose offset for the IMU relative to the based, used to
compute the IMU angular velocity and linear acceleration. The offset
orientation is the identity, and the offset translation is set based on
specified IMU location (0.116m forward, 0.017m upward). Tested only
to verify that the identity rotation and specified translation appear
to work correctly in transforming IMU data.